Renombrado en peliculas y añadido almacenamiento local de superheroes y peliculas
