+++ app/build.gradle.kts
@@ -37,12 +37,18 @@ android {
 
 dependencies {
 
+    //Librerias para la app
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
     implementation(libs.material)
     implementation(libs.androidx.activity)
     implementation(libs.androidx.constraintlayout)
+    implementation(libs.gson.serializer)
+
+    //Librerías para testing
     testImplementation(libs.junit)
+
+    //Librerías para Android-testing
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
 }
