--- app/src/main/java/edu/iesam/dam2024/feature/movies/domain/GetMovieUseCase.kt
@@ -1,8 +1,0 @@
-package edu.iesam.dam2024.feature.movies.domain
-
-class GetMovieUseCase(private val movieRepository: MovieRepository) {
-
-    operator fun invoke(movieId: String): Movie? {
-        return movieRepository.getMovie(movieId)
-    }
-}
