--- app/src/main/java/edu/iesam/dam2024/feature/movies/domain/GetMoviesUseCase.kt
@@ -1,12 +1,0 @@
-package edu.iesam.dam2024.feature.movies.domain
-
-/**
- * Naming= Verbo(Acción)+ Modelo + UseCase
- */
-
-class GetMoviesUseCase(private val movieRepository: MovieRepository) {
-
-    operator fun invoke(): List<Movie> {
-        return movieRepository.getMovies()
-    }
-}
