--- app/src/main/java/edu/iesam/dam2024/feature/movies/domain/Movie.kt
@@ -1,3 +1,0 @@
-package edu.iesam.dam2024.feature.movies.domain
-
-data class Movie(val id: String, val title: String, val poster: String)
