--- app/src/main/java/edu/iesam/dam2024/feature/movies/domain/MovieRepository.kt
@@ -1,8 +1,0 @@
-package edu.iesam.dam2024.feature.movies.domain
-
-interface MovieRepository {
-
-    fun getMovies(): List<Movie>
-    fun getMovie(movieId: String): Movie?
-
-}
