--- app/src/main/java/edu/iesam/dam2024/feature/superhero/domain/Appearance.kt
@@ -1,10 +1,0 @@
-package edu.iesam.dam2024.feature.superhero.domain
-
-data class Appearance(
-    val gender: String,
-    val race: String,
-    val height: List<String>,
-    val weight: List<String>,
-    val eyeColor: String,
-    val hairColor: String
-)
