--- app/src/main/java/edu/iesam/dam2024/feature/superhero/domain/Biography.kt
@@ -1,11 +1,0 @@
-package edu.iesam.dam2024.feature.superhero.domain
-
-data class Biography(
-    val fullName: String,
-    val alterEgos: String,
-    val aliases: List<String>,
-    val placeOfBirth: String,
-    val firstAppearance: String,
-    val publisher: String,
-    val alignment: String
-)
