--- app/src/main/java/edu/iesam/dam2024/feature/superhero/domain/Connections.kt
@@ -1,6 +1,0 @@
-package edu.iesam.dam2024.feature.superhero.domain
-
-data class Connections(
-    val groupAffiliation: String,
-    val relatives: String
-)
