--- app/src/main/java/edu/iesam/dam2024/feature/superhero/domain/GetSuperHeroUseCase.kt
@@ -1,9 +1,0 @@
-package edu.iesam.dam2024.feature.superhero.domain
-
-class GetSuperHeroUseCase(private val superHeroRepository: SuperHeroRepository) {
-
-    operator fun invoke(superHeroId: String): SuperHero? {
-        return superHeroRepository.getSuperHero(superHeroId)
-    }
-
-}
