--- app/src/main/java/edu/iesam/dam2024/feature/superhero/domain/GetSuperHeroesUseCase.kt
@@ -1,9 +1,0 @@
-package edu.iesam.dam2024.feature.superhero.domain
-
-class GetSuperHeroesUseCase(private val superHeroRepository: SuperHeroRepository) {
-
-    operator fun invoke(): List<SuperHero> {
-        return superHeroRepository.getSuperHeroes()
-    }
-
-}
