--- app/src/main/java/edu/iesam/dam2024/feature/superhero/domain/PowerStats.kt
@@ -1,10 +1,0 @@
-package edu.iesam.dam2024.feature.superhero.domain
-
-data class PowerStats(
-    val intelligence: Int,
-    val strength: Int,
-    val speed: Int,
-    val durability: Int,
-    val power: Int,
-    val combat: Int
-)
