--- app/src/main/java/edu/iesam/dam2024/feature/superhero/domain/SuperHeroRepository.kt
@@ -1,8 +1,0 @@
-package edu.iesam.dam2024.feature.superhero.domain
-
-interface SuperHeroRepository {
-
-    fun getSuperHeroes(): List<SuperHero>
-    fun getSuperHero(superHeroId: String): SuperHero?
-
-}
