--- app/src/main/java/edu/iesam/dam2024/feature/superhero/domain/Work.kt
@@ -1,6 +1,0 @@
-package edu.iesam.dam2024.feature.superhero.domain
-
-data class Work(
-    val occupation: String,
-    val base: String
-)
