+++ app/src/main/java/edu/iesam/dam2024/features/movies/domain/GetMoviesUseCase.kt
@@ -1,0 +1,12 @@
+package edu.iesam.dam2024.features.movies.domain
+
+/**
+ * Naming= Verbo(Acción)+ Modelo + UseCase
+ */
+
+class GetMoviesUseCase(private val movieRepository: MovieRepository) {
+
+    operator fun invoke(): List<Movie> {
+        return movieRepository.getMovies()
+    }
+}
