+++ app/src/main/java/edu/iesam/dam2024/features/movies/domain/Movie.kt
@@ -1,0 +1,3 @@
+package edu.iesam.dam2024.features.movies.domain
+
+data class Movie(val id: String, val title: String, val poster: String)
