+++ app/src/main/java/edu/iesam/dam2024/features/movies/domain/MovieRepository.kt
@@ -1,0 +1,8 @@
+package edu.iesam.dam2024.features.movies.domain
+
+interface MovieRepository {
+
+    fun getMovies(): List<Movie>
+    fun getMovie(movieId: String): Movie?
+
+}
