+++ app/src/main/java/edu/iesam/dam2024/features/superhero/domain/Appearance.kt
@@ -1,0 +1,10 @@
+package edu.iesam.dam2024.features.superhero.domain
+
+data class Appearance(
+    val gender: String,
+    val race: String,
+    val height: List<String>,
+    val weight: List<String>,
+    val eyeColor: String,
+    val hairColor: String
+)
