+++ app/src/main/java/edu/iesam/dam2024/features/superhero/domain/Connections.kt
@@ -1,0 +1,6 @@
+package edu.iesam.dam2024.features.superhero.domain
+
+data class Connections(
+    val groupAffiliation: String,
+    val relatives: String
+)
