+++ app/src/main/java/edu/iesam/dam2024/features/superhero/domain/GetSuperHeroUseCase.kt
@@ -1,0 +1,9 @@
+package edu.iesam.dam2024.features.superhero.domain
+
+class GetSuperHeroUseCase(private val superHeroRepository: SuperHeroRepository) {
+
+    operator fun invoke(superHeroId: String): SuperHero? {
+        return superHeroRepository.getSuperHero(superHeroId)
+    }
+
+}
