+++ app/src/main/java/edu/iesam/dam2024/features/superhero/domain/GetSuperHeroesUseCase.kt
@@ -1,0 +1,9 @@
+package edu.iesam.dam2024.features.superhero.domain
+
+class GetSuperHeroesUseCase(private val superHeroRepository: SuperHeroRepository) {
+
+    operator fun invoke(): List<SuperHero> {
+        return superHeroRepository.getSuperHeroes()
+    }
+
+}
