+++ app/src/main/java/edu/iesam/dam2024/features/superhero/domain/PowerStats.kt
@@ -1,0 +1,10 @@
+package edu.iesam.dam2024.features.superhero.domain
+
+data class PowerStats(
+    val intelligence: Int,
+    val strength: Int,
+    val speed: Int,
+    val durability: Int,
+    val power: Int,
+    val combat: Int
+)
