+++ app/src/main/java/edu/iesam/dam2024/features/superhero/domain/SuperHero.kt
@@ -1,0 +1,12 @@
+package edu.iesam.dam2024.features.superhero.domain
+
+data class SuperHero(
+    val id: String,
+    val name: String,
+    val slug: String,
+    val powerStats: PowerStats,
+    val appearance: Appearance,
+    val biography: Biography,
+    val work: Work,
+    val connections: Connections
+)
