+++ app/src/main/java/edu/iesam/dam2024/features/superhero/domain/SuperHeroRepository.kt
@@ -1,0 +1,8 @@
+package edu.iesam.dam2024.features.superhero.domain
+
+interface SuperHeroRepository {
+
+    fun getSuperHeroes(): List<SuperHero>
+    fun getSuperHero(superHeroId: String): SuperHero?
+
+}
