+++ app/src/main/java/edu/iesam/dam2024/features/superhero/domain/Work.kt
@@ -1,0 +1,6 @@
+package edu.iesam.dam2024.features.superhero.domain
+
+data class Work(
+    val occupation: String,
+    val base: String
+)
