+++ app/src/main/java/edu/iesam/dam2024/features/superhero/presentation/SuperHeroFactory.kt
@@ -1,0 +1,18 @@
+package edu.iesam.dam2024.features.superhero.presentation
+
+import edu.iesam.dam2024.features.superhero.data.SuperHeroDataRepository
+import edu.iesam.dam2024.features.superhero.data.remote.SuperHeroMockRemoteDataSource
+import edu.iesam.dam2024.features.superhero.domain.GetSuperHeroUseCase
+import edu.iesam.dam2024.features.superhero.domain.GetSuperHeroesUseCase
+
+class SuperHeroFactory {
+
+    private val superHeroMockRemoteDataSource = SuperHeroMockRemoteDataSource()
+    private val superHeroDataRepository = SuperHeroDataRepository(superHeroMockRemoteDataSource)
+    private val getSuperHeroUseCase = GetSuperHeroUseCase(superHeroDataRepository)
+    private val getSuperHeroesUseCase = GetSuperHeroesUseCase(superHeroDataRepository)
+
+    fun buildViewModel(): SuperHeroViewModel {
+        return SuperHeroViewModel(getSuperHeroesUseCase, getSuperHeroUseCase)
+    }
+}
