+++ app/src/main/java/edu/iesam/dam2024/features/superhero/presentation/SuperHeroViewModel.kt
@@ -1,0 +1,20 @@
+package edu.iesam.dam2024.features.superhero.presentation
+
+import androidx.lifecycle.ViewModel
+import edu.iesam.dam2024.features.superhero.domain.GetSuperHeroUseCase
+import edu.iesam.dam2024.features.superhero.domain.GetSuperHeroesUseCase
+import edu.iesam.dam2024.features.superhero.domain.SuperHero
+
+class SuperHeroViewModel(
+    private val getSuperHeroesUseCase: GetSuperHeroesUseCase,
+    private val getSuperHeroUseCase: GetSuperHeroUseCase
+) : ViewModel() {
+
+    fun viewCreated(): List<SuperHero> {
+        return getSuperHeroesUseCase.invoke()
+    }
+
+    fun itemSelected(superHeroId: String): SuperHero? {
+        return getSuperHeroUseCase.invoke(superHeroId)
+    }
+}
