+++ app/src/main/res/layout/activity_superhero.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".feature.superhero.presentation.SuperHeroActivity">
+    tools:context=".features.superhero.presentation.SuperHeroActivity">
 
     <LinearLayout
         android:id="@+id/layout1_superhero"
